requirements.txt
streamlit
langchain==0.3.*
langchain-community==0.3.*
faiss-cpu
PyPDF2
python-dotenv
requests
numpy
pandas
sentence-transformers
langchain-google-genai